# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ipon Buffer",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/sandbox/reset": {
      "post": {
        "summary": "Reset",
        "operationId": "reset_api_v1_sandbox_reset_post",
        "parameters": [
          {
            "name": "X-Demo-Session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/consents": {
      "post": {
        "summary": "Grant",
        "operationId": "grant_api_v1_consents_post",
        "parameters": [
          {
            "name": "X-Demo-Session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Consents",
        "operationId": "consents_api_v1_consents_get",
        "parameters": [
          {
            "name": "X-Demo-Session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Revoke",
        "operationId": "revoke_api_v1_consents_delete",
        "parameters": [
          {
            "name": "X-Demo-Session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/wallet": {
      "get": {
        "summary": "Wallet",
        "operationId": "wallet_api_v1_wallet_get",
        "parameters": [
          {
            "name": "X-Demo-Session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/summary": {
      "get": {
        "summary": "Get Summary",
        "operationId": "get_summary_api_v1_summary_get",
        "parameters": [
          {
            "name": "X-Demo-Session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plan": {
      "put": {
        "summary": "Put Plan",
        "operationId": "put_plan_api_v1_plan_put",
        "parameters": [
          {
            "name": "X-Demo-Session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recommendation": {
      "post": {
        "summary": "Recommendation",
        "operationId": "recommendation_api_v1_recommendation_post",
        "parameters": [
          {
            "name": "X-Demo-Session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/buffer/transfer": {
      "post": {
        "summary": "Deposit",
        "operationId": "deposit_api_v1_buffer_transfer_post",
        "parameters": [
          {
            "name": "X-Demo-Session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/buffer/withdraw": {
      "post": {
        "summary": "Withdraw",
        "operationId": "withdraw_api_v1_buffer_withdraw_post",
        "parameters": [
          {
            "name": "X-Demo-Session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sandbox/pay-emergency": {
      "post": {
        "summary": "Emergency",
        "operationId": "emergency_api_v1_sandbox_pay_emergency_post",
        "parameters": [
          {
            "name": "X-Demo-Session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConsentInput": {
        "properties": {
          "scopes": {
            "items": {
              "type": "string",
              "enum": [
                "transactions:read",
                "balance:read",
                "buffer:transfer",
                "ai:parse"
              ]
            },
            "type": "array",
            "title": "Scopes"
          },
          "days": {
            "type": "integer",
            "enum": [
              30,
              90,
              180
            ],
            "title": "Days",
            "default": 90
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "scopes"
        ],
        "title": "ConsentInput"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlanInput": {
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Version"
          },
          "confirmed": {
            "type": "boolean",
            "title": "Confirmed"
          },
          "routine_daily_minor": {
            "type": "integer",
            "maximum": 100000000.0,
            "minimum": 0.0,
            "title": "Routine Daily Minor"
          },
          "next_expected_income": {
            "type": "string",
            "format": "date",
            "title": "Next Expected Income"
          },
          "obligations": {
            "items": {
              "$ref": "#/components/schemas/BillInput"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Obligations"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "version",
          "confirmed",
          "routine_daily_minor",
          "next_expected_income",
          "obligations"
        ],
        "title": "PlanInput"
      },
...
    }
  }
}